requirements.txt
six
# Old python needs reqests<=2.18
requests<=2.18.4; python_version<'3'
requests; python_version>='3'
jsonschema==2.5.1
# Old Python needs old PyYAML
PyYAML<3.12; python_version<'3'
# New Python can use newest PyYAML
PyYAML; python_version>='3'
ordereddict; python_version<'2.7'